# docs/ingestion-openapi.yaml
openapi: 3.0.3
info:
  title: Tech Advisor Ingestion API
  version: 1.0.0
  description: >-
    ADMIN-only ingestion operations authenticated with the application's normal bearer JWT.
    The ingestion-demo profile only enables simulated sources; it does not change authentication.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/admin/ingestion/session:
    get:
      summary: Read the authenticated administrator identity
      responses:
        '200':
          description: Authenticated administrator
          content:
            application/json:
              schema:
                type: object
                properties:
                  username: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/admin/ingestion/searchapi/candidates:
    post:
      summary: Find validated SearchAPI products for manual selection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productName]
              properties:
                productName: { type: string, minLength: 1, maxLength: 200 }
      responses:
        '200':
          description: Up to 20 valid provider-ranked identities. Product tokens remain server-only.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [externalProductId, title]
                  properties:
                    externalProductId: { type: string }
                    title: { type: string }
        '400': { $ref: '#/components/responses/Invalid' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { description: SearchAPI rate limited the discovery request }
        '502': { description: SearchAPI returned an unusable response }
        '503': { description: SearchAPI could not be reached }
  /api/admin/ingestion/runs:
    post:
      summary: Request an asynchronous scrape without moving the recurring schedule
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, pattern: '^[A-Za-z0-9_-]{8,128}$' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sources:
                  type: array
                  description: Registered enabled source IDs; absent or empty selects all enabled sources.
                  items: { type: string }
                reason: { type: string, nullable: true, maxLength: 500 }
                productName:
                  type: string
                  nullable: true
                  minLength: 1
                  maxLength: 200
                  description: >-
                    Optional for SearchAPI manual runs. Brand followed by the full model, ignoring repeated
                    whitespace. When present, externalProductId is required. Existing eligible smartphones
                    resolve before admission. An unknown name is rejected because MobileAPI is the sole
                    source of catalogue products. Omission/null retains source-default product selection.
                externalProductId:
                  type: string
                  nullable: true
                  maxLength: 1000
                  description: External ID returned by the candidate endpoint; never a provider product token.
      responses:
        '202':
          description: Durable run admission, or an existing run for the same actor/key/body.
          headers:
            Location:
              schema: { type: string }
              description: Run detail URL
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '400': { $ref: '#/components/responses/Invalid' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Active run (ID in message), or idempotency key reused with a different body.
        '503': { $ref: '#/components/responses/Unavailable' }
    get:
      summary: Paginated run history, newest first
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 0, maximum: 10000, default: 0 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: trigger
          schema: { type: string, enum: [MANUAL, SCHEDULED] }
      responses:
        '200':
          description: Runs (empty array past the last page)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Run' }
        '400': { $ref: '#/components/responses/Invalid' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/admin/ingestion/runs/{id}:
    get:
      summary: Read persisted run state and source results
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Run state, including in-flight counters
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '404': { description: Unknown run ID }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/admin/ingestion/sources:
    get:
      summary: Registered sources and persistent cooldowns
      responses:
        '200':
          description: Available source modules
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    sourceId: { type: string }
                    enabled: { type: boolean }
                    simulation: { type: boolean }
                    nextAllowedAt: { type: string, format: date-time, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/admin/ingestion/schedule:
    get:
      summary: Current schedule and active run
      responses:
        '200':
          description: Persisted schedule (manual runs never move nextScheduledAt)
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled: { type: boolean }
                  intervalDays: { type: integer, enum: [2] }
                  intervalHours: { type: integer, enum: [48] }
                  anchor: { type: string, format: date-time, nullable: true }
                  nextScheduledAt: { type: string, format: date-time, nullable: true }
                  activeRunId: { type: string, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/Unavailable' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Signed account JWT whose role claim and current database account both resolve to ADMIN.
  responses:
    Unauthorized: { description: Authentication required }
    Forbidden: { description: Authenticated account does not have the ADMIN role }
    Invalid: { description: Invalid source selection, request, or pagination }
    Unavailable: { description: Run storage unavailable }
  schemas:
    SourceResult:
      type: object
      properties:
        sourceId: { type: string }
        status: { type: string }
        startedAt: { type: string, format: date-time, nullable: true }
        finishedAt: { type: string, format: date-time, nullable: true }
        processedPayloadCount: { type: integer }
        duplicatePayloadCount: { type: integer }
        rejectedPayloadCount: { type: integer }
        errorCount: { type: integer }
        errorStackCount: { type: integer }
        errors:
          type: array
          description: Sanitized exception type and at most five application stack frames.
          items: { type: string }
    Run:
      type: object
      properties:
        runId: { type: string }
        triggerType: { type: string, enum: [MANUAL, SCHEDULED] }
        requestedBy: { type: string }
        reason: { type: string, nullable: true }
        product:
          type: object
          nullable: true
          description: Existing canonical product plus an optional admin-selected provider identity; included in idempotency comparison and preserved across dispatch/restarts.
          properties:
            productId: { type: integer, format: int64, nullable: true }
            productName: { type: string }
            externalProductId: { type: string, nullable: true }
        idempotencyKey: { type: string, nullable: true }
        status:
          type: string
          enum: [ACCEPTED, RUNNING, SUCCESS, PARTIAL_FAILURE, FAILED, INTERRUPTED, SKIPPED]
        requestedAt: { type: string, format: date-time }
        scheduledFor: { type: string, format: date-time, nullable: true }
        startedAt: { type: string, format: date-time, nullable: true }
        finishedAt: { type: string, format: date-time, nullable: true }
        durationMs: { type: integer, format: int64 }
        latenessMs: { type: integer, format: int64 }
        missedSlots: { type: integer, format: int64 }
        processedPayloadCount: { type: integer }
        duplicatePayloadCount: { type: integer }
        rejectedPayloadCount: { type: integer }
        errorCount: { type: integer }
        errorStackCount: { type: integer }
        skippedSourceCount: { type: integer }
        simulation: { type: boolean }
        sourceIds:
          type: array
          items: { type: string }
        sources:
          type: array
          items: { $ref: '#/components/schemas/SourceResult' }